max/min runs dependent on polynomial degree
Adjusted program so that the program runs to find as many zeroes or max/mins as the biggest coefficient dictates (i.e. if the highest degree is 5, it will run maxmin method and zero method five times). This way there is a dependence between how many times the code runs and the degree of the polynomial